Dockerisation

+++ src/participants/participants.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Participant } from './participant.schema';
 import { Model } from 'mongoose';
@@ -20,13 +20,22 @@ export class ParticipantsService {
     }
 
     async updateParticipant(participantId: string, updateParticipantDto: UpdateParticipantDto): Promise<Participant>{
-        const participant = this.participantModel.findByIdAndUpdate(participantId,updateParticipantDto,{new:true});
-        return participant;
+        const participant = await this.participantModel.findById(participantId);
+        if(! participant){
+            throw new NotFoundException('Participant not found');
+        }
+        return await this.participantModel.findByIdAndUpdate(participantId,updateParticipantDto,{new:true});
+        
     }
 
     async removeParticipant(participantId: string, eventId: string){
-        const participant = await this.participantModel.findByIdAndDelete(participantId);
-        await this.eventService.removeParticipant(eventId,participantId);
-        return participant;
+
+        const participant = await this.participantModel.findById(participantId);
+        if(! participant){
+            throw new NotFoundException('Participant not found');
+        }
+        await this.participantModel.findByIdAndDelete(participantId);
+        return await this.eventService.removeParticipant(eventId,participantId);
+        
     }
 }
